% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{as_r2slides_table}
\alias{as_r2slides_table}
\title{Convert a table object to an r2slides_table}
\usage{
as_r2slides_table(x, ...)
}
\arguments{
\item{x}{A table object. Methods are expected for classes such as \code{gt_tbl}
(from \strong{gt}) and \code{flextable} (from \strong{flextable}).}

\item{...}{Additional arguments passed to methods.}
}
\value{
An \link{r2slides_table} object.
}
\description{
\code{as_r2slides_table()} is a generic function that converts a table object
(e.g. from \strong{gt} or \strong{flextable}) into an \link{r2slides_table} object suitable
for rendering as a Google Slides table element.

Methods for specific table classes should return an \link{r2slides_table} object.
}
