% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_position.R
\name{chart_annotation_1}
\alias{chart_annotation_1}
\title{Testing function for how to use define_relative_transformation_function}
\usage{
chart_annotation_1(position)
}
\arguments{
\item{position}{An object of class `r2slides::slide_position`}
}
\value{
An object of class `r2slides::slide_position`
}
\description{
Testing function for how to use define_relative_transformation_function
}
\examples{
\dontrun{
data |>
  write_gs("Sheet Name") |>
  add_linked_chart(on_slide_url("URL Of Slide"), in_top_left()) |>
  add_text("Title", in_top_left() |> chart_annotation_1())
}
}
