% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_chart_class.R
\name{chart_id}
\alias{chart_id}
\title{Chart ID class}
\usage{
chart_id(
  spreadsheet_id = character(0),
  sheet_id = character(0),
  chart_id = character(0)
)
}
\arguments{
\item{spreadsheet_id}{A single string. The Google Sheets spreadsheet ID.}

\item{sheet_id}{A single string. The sheet tab ID.}

\item{chart_id}{A single string. The chart ID.}
}
\description{
A typed reference to a specific chart within a Google Sheets sheet. Inherits
from \link{sht_id}.
}
