% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_strategy.R
\name{clear_request_buffer}
\alias{clear_request_buffer}
\title{Clear the request buffer}
\usage{
clear_request_buffer()
}
\value{
\code{NULL}, invisibly.
}
\description{
Removes all entries from the internal request buffer. Useful for resetting
state between workflows.
}
\seealso{
\code{\link[=view_request_buffer]{view_request_buffer()}}, \code{\link[=execute_requests]{execute_requests()}}
}
