% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_text.R
\name{combine_style}
\alias{combine_style}
\title{Combine Text Styles}
\usage{
combine_style(..., .priority = c("first", "last", "error"))
}
\arguments{
\item{...}{Text styles to combine}

\item{.priority}{Priority order in which to combine styles}
}
\value{
A text_style object
}
\description{
Combine text styles in a priority order
}
