% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_slide_size.R
\name{correct_slide_size}
\alias{correct_slide_size}
\title{Correct slide sizes}
\usage{
correct_slide_size(
  pos,
  dim = c("width", "height"),
  slide_size = NULL,
  call = rlang::caller_env()
)
}
\arguments{
\item{pos}{A length one numeric vector.}

\item{dim}{One of `"width"` or `"height"`.}

\item{slide_size}{Optional. A list containing slide size specifications for converting. In the form of list(x_height = 7.5, x_width = 13.3, y_height = 5, y_width = 9). `x_` indicates converting from and `y_` indicates converting too. You can override this to convert between custom slide sizes. If left blank it converts from PowerPoint to Google Slides.}

\item{call}{Optional. The execution environment, used for error messages.}
}
\value{
A numeric value of corrected position/size. Will error if `slide_size` is provided
but is not a list.
}
\description{
Google Slides uses a different default slide size than PowerPoint.
In order to make it easy for the user, and for re-usable code when working with both
Google Slides and PowerPoint, this function converts sizes and positions that use the
PowerPoint system to the Google Slides System. You may also override the slide_size argument
to convert between custom slide sizes.
}
