% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation_helpers.R
\name{create_presentation_env_in_global}
\alias{create_presentation_env_in_global}
\title{Create presentation environment}
\usage{
create_presentation_env_in_global()
}
\value{
Creates a `google_presentation` environment in the global environment if one does
not already exist.
}
\description{
Create presentation environment
}
