% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheet_helpers.R
\name{create_spreadsheet_env_in_global}
\alias{create_spreadsheet_env_in_global}
\title{Create spreadsheet environment}
\usage{
create_spreadsheet_env_in_global()
}
\value{
Creates a `google_spreadsheet` environment in the global environment if one does
not already exist.
}
\description{
Create spreadsheet environment
}
