% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_text.R
\name{create_styling_request}
\alias{create_styling_request}
\title{Create a styling request}
\usage{
create_styling_request(
  style_rule,
  text,
  element_id,
  ...,
  call = rlang::caller_env()
)
}
\arguments{
\item{style_rule}{A style_rule object}

\item{text}{A character string of text}

\item{element_id}{A string ID of an existing text element to update. If element_id is `NULL` a new element will be created.}

\item{...}{Data to be avaiable in the data mask where the selector function will be executed. Must be named.}

\item{call}{Optional. Call environment used in error messages.}
}
\value{
A list of Google Slides styling requests
}
\description{
Create a styling request
}
