% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_position.R
\name{define_relative_transformation_function}
\alias{define_relative_transformation_function}
\title{This function takes transformations for top, left, width, and height, and returns a function that applies those transformations to a slide_position object.}
\usage{
define_relative_transformation_function(
  top_transformation = identity,
  left_transformation = identity,
  width_transformation = identity,
  height_transformation = identity
)
}
\arguments{
\item{top_transformation}{A function to apply to the top value, or a scalar numeric to replace it}

\item{left_transformation}{A function to apply to the left value, or a scalar numeric to replace it}

\item{width_transformation}{A function to apply to the width value, or a scalar numeric to replace it}

\item{height_transformation}{A function to apply to the height value, or a scalar numeric to replace it}
}
\value{
A function that applies those transformations to a slide_position object.
}
\description{
This function takes transformations for top, left, width, and height, and returns a function that applies those transformations to a slide_position object.
}
