% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_presentation.R
\name{get_active_presentation}
\alias{get_active_presentation}
\title{Get the currently active presentation}
\usage{
get_active_presentation()
}
\value{
A `presentation` object, or errors if none active
}
\description{
Get the currently active presentation
}
