% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_spreadsheet.R
\name{get_active_spreadsheet}
\alias{get_active_spreadsheet}
\title{Get the currently active spreadsheet}
\usage{
get_active_spreadsheet()
}
\value{
A `spreadsheet` object, or errors if none is active
}
\description{
Get the currently active spreadsheet
}
