% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_chart_class.R
\name{get_chart_id}
\alias{get_chart_id}
\title{Get chart_id from a Google Sheet}
\usage{
get_chart_id(spreadsheet_obj, token = NULL, call = rlang::caller_env())
}
\arguments{
\item{spreadsheet_obj}{A `sht_id` object.}

\item{token}{Optional. An OAuth2 token. The default uses `r2slides_token()`.}

\item{call}{Optional. Call environment used in error messages.}
}
\value{
A `chart_id` object
}
\description{
Retrieves the chart ID from a specified sheet. Errors if 0 or more than 1
chart is found.
}
