% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_strategy.R
\name{get_evaluation_strategy}
\alias{get_evaluation_strategy}
\title{Get the current request evaluation strategy}
\usage{
get_evaluation_strategy()
}
\value{
A string, either \code{"eager"} or \code{"lazy"}.
}
\description{
Returns the value of the \code{r2slides_request_evaluation_strategy} environment
variable, defaulting to \code{"eager"} if unset.
}
\seealso{
\code{\link[=set_evaluation_strategy]{set_evaluation_strategy()}}
}
