% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_generics.R
\name{get_position}
\alias{get_position}
\title{Get the position of an element}
\usage{
get_position(x, ...)
}
\arguments{
\item{x}{An \code{element} object or subclass.}
}
\value{
A \code{slide_position} object. Can be passed directly to \code{add_text()},
\code{add_image()}, etc. to recreate the element at the same location.
}
\description{
Parses the element's affine transform from the raw API node and returns a
\code{slide_position} object encoding top, left, width, height (all in inches),
and rotation (in degrees). Works on any element subclass.

Google Slides stores positions using an affine transform matrix
(\code{scaleX}, \code{scaleY}, \code{shearX}, \code{shearY}, \code{translateX}, \code{translateY}) plus
a nominal size. \code{get_position()} decomposes the matrix to recover the
rendered dimensions and rotation angle.
}
\examples{
if (FALSE) {
  ps <- register_presentation(id = "YOUR_PRESENTATION_ID", set_active = FALSE)
  slide_obj <- on_slide_number(1, ps)
  el <- get_elements(slide_obj, type = "text")[[1]]
  pos <- get_position(el)
  pos@top
  pos@left
}

}
\seealso{
\code{\link[=slide_position]{slide_position()}}, \code{\link[=get_raw]{get_raw()}}
}
