% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_text.R
\name{get_safe_length}
\alias{get_safe_length}
\title{Get the length of an argument}
\usage{
get_safe_length(arg, arg_name = rlang::caller_arg(arg))
}
\arguments{
\item{arg}{An argument to get the length of.}
}
\description{
Get the length of an argument
}
