% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hex_to_rgb}
\alias{hex_to_rgb}
\title{Convert a hex color code to RGB values}
\usage{
hex_to_rgb(hex_color)
}
\arguments{
\item{hex_color}{A single string representing a 6-digit hex color code, with or without a leading "#".}
}
\value{
A list with three components: `red`, `green`, and `blue`, each containing a number between 0 and 1.
Will error if the input is not a valid 6-digit hex color code.
}
\description{
Convert a hex color code to RGB values
}
\keyword{internal}
