% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_position.R
\name{in_qualtrics_title}
\alias{in_qualtrics_title}
\title{Creates a slide_position object with defaults in the title of the slide}
\usage{
in_qualtrics_title(
  revision = Sys.Date(),
  convert_slide_size = FALSE,
  slide_size_old = NULL,
  slide_size = c(5.625, 10)
)
}
\arguments{
\item{revision}{A date object to specify the defaults for qualtrics title as of a given date}

\item{convert_slide_size}{A logical. Optional.}

\item{slide_size_old}{Old Slide size specifications. Optional. Used for conversion between powerpoint and google slides sizes}

\item{slide_size}{New Slide size specifications. Optional.}
}
\description{
Creates a slide_position object with defaults in the title of the slide
}
