% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{in_to_pt}
\alias{in_to_pt}
\title{Convert inches to points}
\usage{
in_to_pt(x)
}
\arguments{
\item{x}{A numeric vector of measurements in inches.}
}
\value{
A numeric vector of measurements in points (1/72 inches).
Errors if non-numeric input is provided.
}
\description{
Google slides API's can only handle EMU or PT's. This function converts inches to PT's so the user can continue to work in inches.
}
\keyword{internal}
