% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_presentation.R
\name{is.presentation}
\alias{is.presentation}
\title{is.presentation}
\usage{
is.presentation(x)
}
\arguments{
\item{x}{An object to check}
}
\value{
\code{TRUE} if \code{x} is a \code{presentation}, \code{FALSE} otherwise
}
\description{
Check if an object is a presentation
}
