% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_selection.R
\name{is.slide}
\alias{is.slide}
\title{Check if an object is a slide}
\usage{
is.slide(x)
}
\arguments{
\item{x}{An object to check}
}
\value{
TRUE if the object is a slide, FALSE otherwise
}
\description{
Check if an object is a slide
}
