% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_position.R
\name{is.slide_position}
\alias{is.slide_position}
\title{Is slide_position}
\usage{
is.slide_position(x)
}
\arguments{
\item{x}{An object to check}
}
\value{
TRUE if the object is a slide_position, FALSE otherwise
}
\description{
Check if an object is a slide_position object
}
