% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_spreadsheet.R
\name{is.spreadsheet}
\alias{is.spreadsheet}
\title{Check if an object is a spreadsheet}
\usage{
is.spreadsheet(x)
}
\arguments{
\item{x}{An object to check}
}
\value{
`TRUE` if `x` is a `spreadsheet`, `FALSE` otherwise
}
\description{
Check if an object is a spreadsheet
}
