% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_class.R
\name{is_element}
\alias{is_element}
\title{Test whether an object is an element}
\usage{
is_element(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
\code{TRUE} if \code{x} inherits from \code{element}, \code{FALSE} otherwise.
}
\description{
Test whether an object is an element
}
