% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_chart_class.R
\name{new_chart_id}
\alias{new_chart_id}
\alias{print.chart_id}
\title{Chart Id class}
\usage{
new_chart_id(gs4_sheet, sheet_id, chart_id, call = rlang::caller_env())

\method{print}{chart_id}(x, ...)
}
\arguments{
\item{gs4_sheet}{A googlesheets4 sheet id, or something coercible to one}

\item{sheet_id}{An integer referencing a sheet id}

\item{chart_id}{An integer referencing a chart id}

\item{call}{Optional. Call environment used in error messages.}

\item{x}{An object of class `chart_id`}

\item{...}{Additional arguments from print}
}
\description{
Chart Id class
}
