% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_presentation.R
\name{new_presentation}
\alias{new_presentation}
\title{Create a new Google Slides presentation}
\usage{
new_presentation(id = NULL, title = "Untitled Presentation", set_active = TRUE)
}
\arguments{
\item{id}{Optional. A single string. The Google Slides presentation ID.}

\item{title}{Optional. A single string. The title of the presentation.}

\item{set_active}{Optional. A logical value indicating whether to set the presentation as the active presentation.}
}
\value{
A Presentation object
}
\description{
Create a new Google Slides presentation
}
