% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_chart_class.R
\name{new_sht_id}
\alias{new_sht_id}
\alias{is_sht_id}
\title{Sheet Id class}
\usage{
new_sht_id(gs4_sheet, sheet_id, call = rlang::caller_env())

is_sht_id(x)
}
\arguments{
\item{gs4_sheet}{A googlesheets4 sheet id, or something coercible to one}

\item{sheet_id}{An integer referencing a sheet id}

\item{call}{Optional. Call environment used in error messages.}

\item{x}{An object of class `sht_id`}

\item{...}{Additional arguments from print}
}
\description{
Sheet Id class
}
