% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_slide.R
\name{new_slide}
\alias{new_slide}
\title{Create a new slide in a Google Slides presentation}
\usage{
new_slide(
  presentation = google_presentation,
  layout = "TITLE_AND_BODY",
  verbose = TRUE,
  master = NULL,
  ...
)
}
\arguments{
\item{presentation}{Optional. A Google Slides presentation object. By default it searches for
an environment named `google_presentation`.}

\item{layout}{Optional. A character string specifying the slide layout. Default is "TITLE_AND_BODY".}

\item{verbose}{Optional. A logical value indicating whether to print status updates.}

\item{master}{Currently unused; must be NULL. Reserved for officer compatibility. (May drop in future)}

\item{...}{Additional arguments reserved for future expansion, currently unused.}
}
\value{
Updates the presentation object with the new slide ID, invisibly.
Will error if layout is not a character string or if master is not NULL.
}
\description{
Create a new slide in a Google Slides presentation
}
