% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_spreadsheet.R
\name{new_spreadsheet}
\alias{new_spreadsheet}
\title{Create a new Google Sheets spreadsheet}
\usage{
new_spreadsheet(title = "Untitled Spreadsheet", set_active = TRUE)
}
\arguments{
\item{title}{Optional. A single string. The title of the spreadsheet.}

\item{set_active}{Optional. A logical value indicating whether to set the
spreadsheet as the active spreadsheet. Default: \code{TRUE}.}
}
\value{
A \code{spreadsheet} object
}
\description{
Create a new Google Sheets spreadsheet
}
