% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_chart_class.R
\name{print.sht_id}
\alias{print.sht_id}
\title{Print a sht_id object}
\usage{
\method{print}{sht_id}(x, ...)
}
\arguments{
\item{x}{The object to print}

\item{...}{Additional arguments}
}
\description{
Print a sht_id object
}
