% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query2}
\alias{query2}
\title{Query Google API's}
\usage{
query2(
  endpoint = "slides.presentations.batchUpdate",
  params = list(),
  body = NULL,
  base = NULL,
  token = NULL,
  call = rlang::caller_env(),
  ...
)
}
\arguments{
\item{endpoint}{A single string indicating the API endpoint.}

\item{params}{Optional. A list of parameters for the API request.}

\item{body}{Optional. A request body.}

\item{base}{Optional. A string indicating the API base service name. (i.e. 'slides', or 'sheets')}

\item{token}{Optional. An OAuth2 token. The default uses `r2slides_token()` to find a token.}

\item{call}{Optional. Call environment used in error messages.}

\item{...}{Additional arguments reserved for future expansion.}
}
\value{
A processed API response. Will error if the endpoint is not recognized
or if required parameters are missing or malformed.
}
\description{
Query Google API's
}
