% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2slides-package.R
\docType{package}
\name{r2slides-package}
\alias{r2slides}
\alias{r2slides-package}
\title{r2slides: Google Slides/Sheets creation from R}
\description{
Create Google Slides from R code. This utilizes the Google Slides and Google Sheets API's to create charts and slides from R code. It is roughly similar to Officer, but with the Google Suite.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Adam-Hubbs/r2slides}
  \item Report bugs at \url{https://github.com/Adam-Hubbs/r2slides/issues}
}

}
\author{
\strong{Maintainer}: Adam Hubbs \email{adam@y2analytics.com}

}
\keyword{internal}
