% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{r2slides_deauth}
\alias{r2slides_deauth}
\title{Clear current token}
\usage{
r2slides_deauth()
}
\description{
If you ever get an error saying client error: (401) UNAUTHENTICATED with a warning message of `Unable to refresh token: invalid_grant`,
then run this function, and restart the R Session. It should restart the auth process next time you need it.
}
