% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{r2slides_default_client}
\alias{r2slides_default_client}
\title{Get the default OAuth client}
\usage{
r2slides_default_client(path = NULL, name = NULL)
}
\value{
A gargle OAuth client.
}
\description{
Get the default OAuth client
}
\keyword{internal}
