% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{r2slides_get_key_from_json}
\alias{r2slides_get_key_from_json}
\title{Get an OAuth client from a JSON file}
\usage{
r2slides_get_key_from_json(path = NULL, name = NULL)
}
\arguments{
\item{path}{Optional. A single string specifying the path to a JSON credentials file.}

\item{name}{Optional. A single string specifying the client name.}
}
\value{
A gargle OAuth client.
}
\description{
Get an OAuth client from a JSON file
}
\keyword{internal}
