% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{r2slides_token}
\alias{r2slides_token}
\title{Get authentication token}
\usage{
r2slides_token()
}
\value{
An authentication configuration or `NULL` if authentication is not active.
}
\description{
Gets a Oauth Token. If it is not currently set, it called r2slides_auth for authentication.
}
\keyword{internal}
