% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_spreadsheet.R
\name{register_spreadsheet}
\alias{register_spreadsheet}
\title{Register an existing Google Sheets spreadsheet}
\usage{
register_spreadsheet(id, set_active = TRUE)
}
\arguments{
\item{id}{A spreadsheet ID string or Google Sheets URL.}

\item{set_active}{Optional. A logical value indicating whether to set the
spreadsheet as the active spreadsheet. Default: \code{TRUE}.}
}
\value{
A \code{spreadsheet} object
}
\description{
Registers an existing Google Sheets spreadsheet so you can use it in r2slides.
Supply a spreadsheet ID string or a Google Sheets URL.
}
