% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_spreadsheet.R
\name{register_spreadsheet2}
\alias{register_spreadsheet2}
\title{Register a spreadsheet (using gs4 s3 class)}
\usage{
register_spreadsheet2(id, local = FALSE, call = rlang::caller_env())
}
\arguments{
\item{id}{A length one character, sheets_id, or dribble object.}

\item{local}{Whether to assign this or let the package automatically handle this in a special environment.}

\item{call}{The calling environment.}
}
\value{
A sheets_id object
}
\description{
This function registers a spreadsheet in a new and hopefully simpler way using the sheets_id S3 class from the googlesheets4 package.
}
