% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{send_to_front}
\alias{send_to_front}
\alias{send_to_back}
\alias{send_forward}
\alias{send_backward}
\title{Change the Z-order of a page element}
\usage{
send_to_front(element)

send_to_back(element)

send_forward(element)

send_backward(element)
}
\arguments{
\item{element}{An object of class `element`.}
}
\value{
The `element` object, invisibly.
}
\description{
These four functions adjust the stacking order of a page element on its
slide using the Google Slides
[UpdatePageElementsZOrderRequest](https://developers.google.com/workspace/slides/api/reference/rest/v1/presentations/request#updatepageelementszorderrequest).
}
