% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_chart_class.R
\name{sht_id}
\alias{sht_id}
\alias{is_sht_id}
\title{Sheet ID class}
\usage{
sht_id(spreadsheet_id = character(0), sheet_id = character(0))

is_sht_id(x)
}
\arguments{
\item{spreadsheet_id}{A single string. The Google Sheets spreadsheet ID.}

\item{sheet_id}{A single string. The sheet tab ID.}

\item{x}{An object to test}
}
\description{
A typed reference to a specific sheet (tab) within a Google Sheets spreadsheet.
}
