% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_selection.R
\name{slide}
\alias{slide}
\title{Slide Object}
\usage{
slide(presentation, slide_id)
}
\arguments{
\item{presentation}{A presentation object}

\item{slide_id}{A slide ID}
}
\description{
An object representing a Google Slides slide/presentation combo
}
