% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_spreadsheet.R
\name{spreadsheet}
\alias{spreadsheet}
\title{R6 Class for Google Sheets Spreadsheets}
\description{
An R6 class to represent a Google Sheets spreadsheet. Use [new_spreadsheet()]
or [register_spreadsheet()] to create instances.
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{spreadsheet_id}}{The Google Sheets spreadsheet ID}

    \item{\code{title}}{The title of the spreadsheet}

    \item{\code{sheets}}{A tibble of sheet names and IDs}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-spreadsheet-initialize}{\code{spreadsheet$new()}}
    \item \href{#method-spreadsheet-set_active}{\code{spreadsheet$set_active()}}
    \item \href{#method-spreadsheet-set_not_active}{\code{spreadsheet$set_not_active()}}
    \item \href{#method-spreadsheet-is_active}{\code{spreadsheet$is_active()}}
    \item \href{#method-spreadsheet-get_url}{\code{spreadsheet$get_url()}}
    \item \href{#method-spreadsheet-browse}{\code{spreadsheet$browse()}}
    \item \href{#method-spreadsheet-print}{\code{spreadsheet$print()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spreadsheet-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-spreadsheet-initialize}{}}}
\subsection{\code{spreadsheet$new()}}{
  Create or open a spreadsheet
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{spreadsheet$new(id = NULL, title = "Untitled Spreadsheet", set_active = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{id}}{Spreadsheet ID, URL, or NULL to create new}
      \item{\code{title}}{Title for new spreadsheet (only used if id is NULL)}
      \item{\code{set_active}}{Whether to make this the active spreadsheet}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A new `spreadsheet` object
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spreadsheet-set_active"></a>}}
\if{latex}{\out{\hypertarget{method-spreadsheet-set_active}{}}}
\subsection{\code{spreadsheet$set_active()}}{
  Set this spreadsheet as the active one
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{spreadsheet$set_active()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Self, invisibly
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spreadsheet-set_not_active"></a>}}
\if{latex}{\out{\hypertarget{method-spreadsheet-set_not_active}{}}}
\subsection{\code{spreadsheet$set_not_active()}}{
  Remove this spreadsheet as the active one
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{spreadsheet$set_not_active()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Self, invisibly
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spreadsheet-is_active"></a>}}
\if{latex}{\out{\hypertarget{method-spreadsheet-is_active}{}}}
\subsection{\code{spreadsheet$is_active()}}{
  Check if this is the active spreadsheet
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{spreadsheet$is_active()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Logical
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spreadsheet-get_url"></a>}}
\if{latex}{\out{\hypertarget{method-spreadsheet-get_url}{}}}
\subsection{\code{spreadsheet$get_url()}}{
  Get the Google Sheets URL for this spreadsheet
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{spreadsheet$get_url()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Character URL or NULL
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spreadsheet-browse"></a>}}
\if{latex}{\out{\hypertarget{method-spreadsheet-browse}{}}}
\subsection{\code{spreadsheet$browse()}}{
  Open the spreadsheet in a browser
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{spreadsheet$browse()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Self, invisibly
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spreadsheet-print"></a>}}
\if{latex}{\out{\hypertarget{method-spreadsheet-print}{}}}
\subsection{\code{spreadsheet$print()}}{
  Print method for spreadsheet objects
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{spreadsheet$print(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{Additional arguments (unused)}
    }
    \if{html}{\out{</div>}}
  }
}

}
