% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_text.R
\name{style_text}
\alias{style_text}
\title{Apply text styling to an existing text element}
\usage{
style_text(slide_obj, element_id, style, token = NULL, call = NULL)
}
\arguments{
\item{slide_obj}{A Google Slides slide object.}

\item{element_id}{The object ID of the text element.}

\item{style}{A list of text styling properties.}

\item{token}{Optional. An OAuth2 token. The default uses `r2slides_token()` to find a token.}

\item{call}{Optional. Call environment used in error messages.}
}
\description{
Apply text styling to an existing text element
}
