% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_text.R
\name{style_text_if}
\alias{style_text_if}
\title{Apply text styling to all elements on a slide that match the selector function
Can be used to apply to every text element containing a +, every element containing a -, only the text parts of all elements, etc.}
\usage{
style_text_if(selector, style)
}
\arguments{
\item{selector}{A function that takes a text element and returns TRUE if it should be styled.}

\item{style}{A list of text styling properties.}
}
\description{
Apply text styling to all elements on a slide that match the selector function
Can be used to apply to every text element containing a +, every element containing a -, only the text parts of all elements, etc.
}
