% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_cells.R
\name{table_cell}
\alias{table_cell}
\title{A single cell in an r2slides_table}
\usage{
table_cell(
  row_index = integer(0),
  col_index = integer(0),
  text = NULL,
  style = NULL
)
}
\arguments{
\item{row_index}{0-based integer row index.}

\item{col_index}{0-based integer column index.}

\item{text}{Character string with the cell's text content, or `NULL`.}

\item{style}{A [cell_style] object, or `NULL`.}
}
\description{
A single cell in an r2slides_table
}
