% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation_helpers.R
\name{validatePresentation}
\alias{validatePresentation}
\title{Validate presentation environment}
\usage{
validatePresentation(presentation)
}
\arguments{
\item{presentation}{A single string naming a presentation environment.}
}
\value{
Nothing. The function will error if the specified environment doesn't
exist in \code{.GlobalEnv} or if it's missing any of the required objects
(\code{presentation_id}, \code{slide_ids}, or \code{current_slide_id}).
}
\description{
Validate presentation environment
}
\keyword{internal}
