% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_chart.R
\name{add_linked_chart}
\alias{add_linked_chart}
\title{Add a linked Google Sheets chart to a Google Slides presentation}
\usage{
add_linked_chart(
  chart_obj,
  slide_obj,
  position,
  linked = TRUE,
  token = NULL,
  call = rlang::caller_env()
)
}
\arguments{
\item{chart_obj}{A Google Sheets chart object}

\item{slide_obj}{A Google Slides slide object}

\item{position}{An object of class `r2slides::slide_position`}

\item{linked}{Optional. A logical indicating whether the chart should be linked. Default: `TRUE`.}

\item{token}{Optional. An OAuth2 token. The default uses `r2slides_token()` to find a token.}

\item{call}{Optional. Call environment used in error messages.}
}
\value{
The Google Slides slide object (invisibly)
}
\description{
Add a linked Google Sheets chart to a Google Slides presentation
}
