% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_text.R
\name{add_text}
\alias{add_text}
\title{Add or update text in a Google Slides presentation}
\usage{
add_text(
  text,
  presentation = google_presentation,
  element_id = NULL,
  left = 1,
  top = 1,
  width = 2,
  height = 0.5,
  text_style = NULL,
  verbose = TRUE,
  report_style = c("qualtrics", "municipal", "y2"),
  bg_color = NULL,
  convert_slide_size = TRUE,
  slide_size = NULL,
  ...
)
}
\arguments{
\item{text}{A character string of text to add.}

\item{presentation}{Optional. A Google Slides presentation object. By default it searches for
an environment named `google_presentation`.}

\item{element_id}{Optional. A string ID of an existing text element to update. If element_id is `NULL` a new element will be created.}

\item{left}{Optional. A numeric position from left edge in inches. Default: 1.}

\item{top}{Optional. A numeric position from top edge in inches. Default: 1.}

\item{width}{Optional. A numeric width in inches. Default: 2.}

\item{height}{Optional. A numeric height in inches. Default: 0.5.}

\item{text_style}{Optional. A list of text styling properties.}

\item{verbose}{Optional. A logical indicating whether to print API responses. Default: TRUE.}

\item{report_style}{One of `"qualtrics"`, `"municipal"`, or `"y2"`.}

\item{bg_color}{Optional. A list specifying background color (RGB).}

\item{convert_slide_size}{Optional. A logical indicating whether to convert dimensions from PowerPoint into Google Slides. If `TRUE`, then positioning would work identically to a PowerPoint presentation. If `FALSE` it uses the raw inches directly. Default: TRUE.}

\item{slide_size}{Optional. A list containing slide size specifications for converting. In the form of list(x_height = 7.5, x_width = 13.3, y_height = 5, y_width = 9). `x_` indicates converting from and `y_` indicates converting too. You can override this to convert between custom slide sizes. If left blank it converts from PowerPoint to Google Slides.}

\item{...}{Additional arguments reserved for future expansion.}
}
\value{
The element ID invisibly. Will error if no slide is currently selected or if
text is not a character string.
}
\description{
Add or update text in a Google Slides presentation
}
