% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_text.R
\name{add_text}
\alias{add_text}
\title{Add or update text in a Google Slides presentation}
\usage{
add_text(
  slide_obj,
  text,
  position,
  element_id = NULL,
  text_style = NULL,
  verbose = TRUE,
  bg_color = NULL,
  token = NULL,
  call = rlang::caller_env(),
  ...
)
}
\arguments{
\item{slide_obj}{A Google Slides slide object.}

\item{text}{A character string of text to add.}

\item{position}{An object of class `r2slides::slide_position`}

\item{element_id}{Optional. A string ID of an existing text element to update. If element_id is `NULL` a new element will be created.}

\item{text_style}{Optional. A list of text styling properties.}

\item{verbose}{Optional. A logical indicating whether to print API responses. Default: TRUE.}

\item{bg_color}{Optional. A list specifying background color (RGB).}

\item{token}{Optional. An OAuth2 token. The default uses `r2slides_token()` to find a token.}

\item{call}{Optional. Call environment used in error messages.}

\item{...}{Additional arguments reserved for future expansion.}
}
\value{
The Google Slides slide object (invisibly).
}
\description{
Add or update text in a Google Slides presentation
}
