% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_text.R
\name{add_text_multi}
\alias{add_text_multi}
\title{Add or update text in a Google Slides presentation}
\usage{
add_text_multi(
  slide_obj,
  text,
  position,
  position_base = NULL,
  element_id = NULL,
  text_style = NULL,
  verbose = TRUE,
  token = NULL,
  pass_strategy = c("one", "all"),
  debug = FALSE,
  ...
)
}
\arguments{
\item{slide_obj}{A Google Slides slide object.}

\item{text}{A vector of character strings to add.}

\item{position}{A vector of objects of class `r2slides::slide_position`}

\item{position_base}{A vector of objects of class `r2slides::slide_position`}

\item{element_id}{Optional. A vector of string IDs of an existing text element to update. If element_id is `NULL` a new element will be created.}

\item{text_style}{Optional. A vector of text_style or style_rule objects.}

\item{verbose}{Optional. A logical indicating whether to print API responses. Default: TRUE.}

\item{token}{Optional. An OAuth2 token. The default uses `r2slides_token()` to find a token.}

\item{pass_strategy}{Optional. A strategy to pass additional values to style_rule objects.}

\item{debug}{Optional. A logical indicating whether to print debug messages. Default: FALSE.}

\item{...}{Additional values available to style_rule objects.}
}
\value{
The Google Slides slide object (invisibly).
}
\description{
Add or update text in a Google Slides presentation
}
