% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{as_r2slides_table}
\alias{as_r2slides_table}
\title{Convert a table object to an r2slides_table}
\usage{
as_r2slides_table(x, ...)
}
\arguments{
\item{x}{A table object. Methods are expected for classes such as `gt_tbl`
(from **gt**) and `flextable` (from **flextable**).}

\item{...}{Additional arguments passed to methods.}
}
\value{
An [r2slides_table] object.
}
\description{
`as_r2slides_table()` is a generic function that converts a table object
(e.g. from **gt** or **flextable**) into an [r2slides_table] object suitable
for rendering as a Google Slides table element.

Methods for specific table classes should return an [r2slides_table] object.
}
