% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_class.R
\name{element}
\alias{element}
\alias{text_element}
\title{Element Object}
\usage{
element(presentation, slide, element_id)

text_element(presentation, slide, element_id)
}
\arguments{
\item{presentation}{A presentation object}

\item{slide}{A slide object}

\item{element_id}{An element ID}
}
\description{
An object representing a Google Slides element/slide/presentation combo
}
