% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_class.R
\name{element}
\alias{element}
\alias{text_element}
\title{Element Object}
\usage{
element(presentation = NULL, slide = NULL, element_id = character(0))

text_element(presentation = NULL, slide = NULL, element_id = character(0))
}
\arguments{
\item{presentation}{A presentation object}

\item{slide}{A slide object}

\item{element_id}{An element ID}
}
\description{
An object representing a Google Slides element/slide/presentation combo
}
