% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_class.R
\name{get_all_elements}
\alias{get_all_elements}
\title{Get all elements in a slide}
\usage{
get_all_elements(slide)
}
\arguments{
\item{slide}{A slide object}
}
\value{
A list of element objects
}
\description{
Get all elements in a slide
}
