% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_chart_class.R
\name{get_chart_id}
\alias{get_chart_id}
\title{Get chart_id from a Google Sheet}
\usage{
get_chart_id(spreadsheet_obj, token = NULL, call = rlang::caller_env())
}
\arguments{
\item{spreadsheet_obj}{A list of spreadsheet_id (string) and sheet_id (integer).}

\item{token}{Optional. An OAuth2 token. The default uses `r2slides_token()` to find a token.}

\item{call}{Optional. Call environment used in error messages.}
}
\value{
A list of the spreadsheet_id (string) the sheet_id (integer) and chart_id (integer)
}
\description{
Retrieves Sheet object from a specified sheet. Errors is 0 or more than 1 chart is found.
}
