% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_strategy.R
\name{get_evaluation_strategy}
\alias{get_evaluation_strategy}
\title{Get the current request evaluation strategy}
\usage{
get_evaluation_strategy()
}
\value{
A string, either `"eager"` or `"lazy"`.
}
\description{
Returns the value of the `r2slides_request_evaluation_strategy` environment
variable, defaulting to `"eager"` if unset.
}
\seealso{
[set_evaluation_strategy()]
}
