% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_presentation.R
\name{new_presentation}
\alias{new_presentation}
\title{Create a new Google Slides presentation}
\usage{
new_presentation(name, location = NULL, verbose = TRUE, overwrite = FALSE, ...)
}
\arguments{
\item{name}{A single string specifying the title of the presentation.}

\item{location}{Optional. A folder location for where to create the presentation. Defaults to your My Drive Home.}

\item{verbose}{Optional. A logical indicating whether to print status updates. (DEV)}

\item{overwrite}{Optional. A logical indicating whether to overwrite existing presentations. Defaults to FALSE. (TRUE for testing purposes)}

\item{...}{Additional arguments reserved for future expansion}
}
\value{
Creates a new Google Slides presentation and modifies the global environment.
Returns invisibly.
}
\description{
Create a new Google Slides presentation
}
