% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_selection.R
\name{on_slide_number}
\alias{on_slide_number}
\alias{on_current_slide}
\alias{on_newest_slide}
\alias{on_slide_id}
\alias{on_slide_url}
\title{Specify a slide by ID or URL}
\usage{
on_slide_number(n)

on_current_slide()

on_newest_slide()

on_slide_id(id)

on_slide_url(id)
}
\arguments{
\item{n}{A numeric slide ID}

\item{id}{A slide identifier. Can be either:
- A numeric slide ID
- A string containing a numeric slide ID
- A full Google Slides URL}
}
\value{
A slide object

When a URL is provided, the function validates that the presentation ID
extracted from the URL matches the currently registered presentation.
}
\description{
Creates a slide reference object that can be used to target a specific slide
in a Google Slides presentation. Accepts either a slide ID (numeric) or a
full Google Slides URL.
}
\examples{
\dontrun{
# Using a numeric slide ID
slide_ref <- on_slide_id(6)
}

}
