% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_selection.R
\name{on_slide_id}
\alias{on_slide_id}
\alias{on_slide_url}
\alias{on_slide_number}
\alias{on_slide_after}
\title{Create a slide object}
\usage{
on_slide_id(id, ps)

on_slide_url(url, ps)

on_slide_number(n, ps)

on_slide_after(slide, offset = 1, ps)
}
\arguments{
\item{id}{A slide identifier. Can be either:
- A numeric slide ID
- A string containing slide ID
- A full Google Slides URL}

\item{ps}{A presentation object}

\item{url}{A URL pointing to a Google Slides slide}

\item{n}{A numeric slide ID}

\item{slide}{A slide object}

\item{offset}{A position integer - the number of slides after the reference slide. Can be negative to return slides before the reference slide}
}
\value{
A slide object
}
\description{
Creates a slide reference object that can be used to target a specific slide
in a Google Slides presentation. Accepts either a slide ID or a
full Google Slides URL.
}
\examples{
\dontrun{
# Using slide ID
slide_ref <- on_slide_id('f82nannfsl_0')
}

}
