% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-peek.R
\name{peek}
\alias{peek}
\title{Peek at a slide}
\usage{
peek(x, ...)
}
\arguments{
\item{x}{A \link{slide} object.}

\item{...}{Ignored; present for S7 generic compatibility.}
}
\description{
Displays a thumbnail of the current slide
}
