% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{r2slides_table}
\alias{r2slides_table}
\title{Table Objects}
\usage{
r2slides_table(
  cells = list(),
  n_rows = integer(0),
  n_cols = integer(0),
  col_widths = NULL,
  row_heights = NULL,
  header_rows = 0L
)
}
\description{
Intermediate representation for various table back-ends
}
