% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_presentation.R
\name{register_presentation}
\alias{register_presentation}
\title{Register a Google Slides presentation}
\usage{
register_presentation(
  id = NULL,
  title = "Untitled Presentation",
  set_active = TRUE
)
}
\arguments{
\item{id}{Optional. A single string. The Google Slides presentation ID.}

\item{title}{Optional. A single string. The title of the presentation.}

\item{set_active}{Optional. A logical value indicating whether to set the presentation as the active presentation.}
}
\value{
A Presentation object
}
\description{
Registers an already created google slides presentation so you can modify it in R.
The user supplies either a name, url, or presentation_id. Supplying multiple will cause
the function to error. If multiple presentation match a given identifier (only possible for name)
the function presents you with a dribble (see the `googledrive` package) and ask which one you want.
}
