% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_position.R
\name{slide_position}
\alias{slide_position}
\title{Slide Position Object}
\usage{
slide_position(
  top,
  left,
  width,
  height,
  convert_slide_size = FALSE,
  slide_size_old = NULL,
  slide_size = c(5.625, 10)
)
}
\arguments{
\item{top}{The position in inches from the top of the slide}

\item{left}{The position in inches from the left of the slide}

\item{width}{The width of the object in inches}

\item{height}{The height of the object in inches}

\item{convert_slide_size}{A logical. Optional. Converts the position and size to Google Slides from a PowerPoint (or custom) size if TRUE.}

\item{slide_size_old}{Old Slide size specifications. Optional. Used for conversion between powerpoint and google slides sizes.}

\item{slide_size}{New Slide size specifications. Optional.}
}
\description{
A system for working with Google Slides positions and sizes.
}
