% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data.R
\name{write_gs}
\alias{write_gs}
\title{Writes data to a Google Sheet}
\usage{
write_gs(data, sheet)
}
\arguments{
\item{data}{A data frame to write to the Google Sheet.}

\item{sheet}{The name of the sheet to write to.}
}
\value{
A list of spreadsheet_id and sheet_id
}
\description{
Writes data to a Google Sheet
}
